% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotBasisVectors.R
\name{amplitude_plots}
\alias{amplitude_plots}
\title{Plot amplitudes of each variable about the basis vectors}
\usage{
amplitude_plots(cum.var, pca.var.coord)
}
\arguments{
\item{cum.var}{The product identified using the cum_var function}

\item{pca.var.coord}{Consults the PCA object created using the FactoMineR package for the desired PCA basis vectors}
}
\value{
A series of plots depicting the amplitude of each variable about each PCA basis vector. Each PCA basis vectors is plotted separately, and each variable is represented along the x-axis. The user should use these plots to identify variables that might interact to create new features (e.g. such as indices  or sums of variables).
}
\description{
Plots the amplitude of the variables about each PCA basis vector identified using the cum_var function
}
