% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap_phenotype.R
\name{bootstrap_phenotype}
\alias{bootstrap_phenotype}
\title{Perform a bootstrap analysis across groups and within each feature, then visualize results in a phenotype.}
\usage{
bootstrap_phenotype(
  features_df_row = features_df_row,
  condition_list = condition_list,
  group_label = group_label,
  reference_group = reference_group
)
}
\arguments{
\item{features_df_row}{A dataframe that contains all features to be analyszed with bootstrap comparison. Specify row names for each observation must contain the experimental group.}

\item{condition_list}{A list of factors indicating the group identifiers as they appear in the data frame (row names)}

\item{group_label}{X-axis title.}

\item{reference_group}{A character string of the reference group as it appears in the row name. This indicates the group against which all significance comparisons are made.}
}
\value{
Plots significance phenotype for each group, where colored groups are significantly different than the reference group. Red is significant and greater than normal, blue is significant and less than the reference group. Stores object sig.phen in your Global Environment. sig.phen contains ggplot visualization of each comparison.
}
\description{
Bootstrap comparison between groups and visualizes the significantly different features in a plasticity phenotype.
}
\examples{
bootstrap_phenotype(
    features_df_row = NewFeatures[,ordered_nf_colnames],
    condition_list = c('  normal  ',
                       '  1wk MD  ',
                       '  fluoxetine + 1wk MD  ',
                       '  fluoxetine  '),
    reference_group = '  normal  ',
    group_label = "\nRearing Conditions"
    )
}
