% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IdentifyBasisVectors.R
\name{cum_var}
\alias{cum_var}
\title{Identify PCA basis vectors that describe a thresholdedd value of cumulative variance}
\usage{
cum_var(pca.eig.3, thresh)
}
\arguments{
\item{pca.eig.3}{Consults the pca.eig.3 product contained in the PCA output from the FactoMineR package}

\item{thresh}{Assign the threshold value of cumulative variance captured by sequential PCA basis vectors}
}
\value{
Only those PCA basis vectors that contain the set threshold value of cumulative variance
}
\description{
Identifies and stores the PCA basis vectors that contain the desired amount of cumulative variance
}
