% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenotype_boxplots.R
\name{phenotype_boxplots}
\alias{phenotype_boxplots}
\title{Create colour-coded boxplots for each feature across every group}
\usage{
phenotype_boxplots(
  feature_df = feature_df,
  phenotype_cols = phenotype.cols,
  group_label = group_label
)
}
\arguments{
\item{feature_df}{The dataframe of features used to create phenotypes, containing group identifiers and the values ofr each feature.}

\item{phenotype_cols}{Taken from the plasticity_phenotype function; dataframe containing hexadecimal color-code for each feature.}

\item{group_label}{X-axis title.}
}
\value{
A series of boxplots depicting the expression of each feature across groups, color-code matches phenotypes.
}
\description{
Plot individual boxplots for each feature, with color codes corresponding to the output of the plasticity_phenotype function.
}
