% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amplitude_plots.R
\name{amplitude_plots}
\alias{amplitude_plots}
\title{Plot amplitudes of each variable for a specified number of PCA basis vectors}
\usage{
amplitude_plots(cum.var, pca.var.coord)
}
\arguments{
\item{cum.var}{The positive integer value outputted by the cum_var function (can also be a user-assigned positive integer ranging from 0 – the total number of variables within the dataset).}

\item{pca.var.coord}{Taken from the PCA object created using the FactoMineR package. Used to extract the vector for each variable.}
}
\value{
A series of plots depicting the amplitude about a specified number of PCA basis vectors. Each PCA basis vectors is plotted separately, and each variable is represented along the x-axis. The user should use these plots to identify variables that might interact to create new features (e.g. such as indices  or sums of variables). The features identified after using the amplitude plots will be saved as a dataframe, which will serve as the input for feature_matrix.
}
\description{
Plots the amplitude of each variable for the specified number of PCA basis vectors identified using the cum_var function.  These plots are used to identify novel plasticity features by comparing the amplitudes of each variable about te basis vectors (e.g. to create variable sums/indices).
}
\examples{
amplitude_plots(
     cum.var = 4,
     pca.var.coord = pca.scaled$var$coord
     )
}
