% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FeatureMatrix.R
\name{feature_matrix}
\alias{feature_matrix}
\title{Creates correlation heatmap between PCA basis vectors and feature matrix}
\usage{
feature_matrix(corr.scores.pval, corr.scores.rval, thresh)
}
\arguments{
\item{corr.scores.pval}{The name of the matrix containing bonferroni adjusted p-values}

\item{corr.scores.rval}{The name of the matrix containing the correlation coefficients}

\item{thresh}{The significance threshold value, which identifies the significance value of Pearson's R correlations to be inclded in the matrix}
}
\value{
Correlation matrix plotted as a heatmap, with filled cells representing only the significantly correlated features
}
\description{
Creates a correlation matrix between the identified PCA basis vectors (using cum_var) and the matrix of features (assigned by user), then plots only the significant correlations in a heatmap
}
