% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlasticityPhenotype.R
\name{plasticity_phenotype}
\alias{plasticity_phenotype}
\title{Plot plasticity phenotypes}
\usage{
plasticity_phenotype(
  means = means,
  maxes = maxes,
  mins = mins,
  grouping_mean_column = grouping_mean_column,
  first_mean_sums_column = first_mean_sums_column,
  last_mean_sums_column = last_mean_sums_column,
  first_mean_indices_column = first_mean_indices_column,
  first_max_sums_vector = first_max_sums_vector,
  last_max_sums_vector = last_max_sums_vector,
  grouping_var = grouping_var
)
}
\arguments{
\item{means}{Data frame of group minimum values}

\item{maxes}{Data frame of group maximum values}

\item{mins}{Dataframe of group minimum values}

\item{grouping_mean_column}{Numerical value; The data frame column that coantins grouping variables (e.g. 1)}

\item{first_mean_sums_column}{Numerical value; The first data frame column that contains feature sums (plotted as white-grey-black). If data frame columns 2:4 contain feature sums, set to 2.}

\item{last_mean_sums_column}{Numerical value; The last data frame column that contains feature sums (plotted as white-grey-black). If data frame columns 2:4 contain feature sums, set to 4.}

\item{first_mean_indices_column}{Numerical value; The first data frame column that contains feature indices (plotted as green-yellow-red). If data frame columns 3:7 contain feature indices, set to 3.}
}
\value{
Plots phenotype for each group
}
\description{
Plots a series of stacked, colour-coded horizontal bars for each group. Color of each bar represents the group mean for the designated feature. Produces a phenotype for each group.
}
